% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixFun.R
\name{createAMatrix}
\alias{createAMatrix}
\title{A Matrix Function}
\usage{
createAMatrix(inputNum, numRows, numCols)
}
\arguments{
\item{inputNum}{what you want in the matrix}

\item{numRows}{number of rows}

\item{numCols}{number of columns}
}
\description{
This function makes matrix... pretty straight forward
}
\examples{
createAMatrix(inputNum = 1, numRows = 2, numCols = 4)
}

