% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMatrix.R
\docType{data}
\name{testMatrix}
\alias{testMatrix}
\title{a matrix of 1}
\format{a matrix with 4 rows and 4 columns
\desribe{
\item{V1}{column1}
\item{V2}{column2}
\item{V3}{column3}
\item{V4}{column4}
}}
\usage{
testMatrix
}
\description{
a dataset containing 1's
}
\keyword{datasets}

